Make POSOrderListModel the single surface for the POS refund flow
POSOrderListModel owns both controllers and is the only thing the orders
screens talk to. It forwards refund state and actions, exposes
refundActionAvailability for the selected order, resets the refund flow on
selectOrder, and refreshes the refunded order after processRefund, logging
a failed refresh instead of swallowing it.

POSRefundControllerProtocol is @MainActor, matching the class and the
sibling POSRefundSubmissionProcessing, so mocks and previews run under the
same isolation as production. order and preparation are private again.

selectOrder moves off POSOrderListControllerProtocol onto a separate
POSOrderSelectionHandling protocol that only the model holds, so views
cannot change the selection without resetting the refund flow.

POSOrderListController loses every refund forward and its submission
processor. It keeps the order list, the selected order, and the refund
details for the order details rows. The model's role is documented in the
module README.

POSOrderListControllerTests splits into POSRefundControllerTests and
POSOrderListModelTests. Shared builders live in enum POSOrderTestFactory,
MockPOSRefundSubmissionProcessor moves to Mocks/, and MockPOSRefundController
stubs instead of re-implementing the selection logic.
